Add function create_new_pdf, fix function p_cleaner
Add last function necessary for the program to work, the one that creates the bilingual PDF afte all processing.
Fix function that cleans the paragraphs of useless texts (URL, nummerical paragraphs), and weird lefover special characters.

OBS:
TO DO: When we use Google with translators it gives a warning that some function worked. Need to supress this warning.